refactor: profile tab의 데이터 파싱 로직들을 service 레이어로 분리 (#126)
