+++ src/service/profile/getProblemsCnt.ts
@@ -1,0 +1,8 @@
+import { ProblemsCntType } from '@src/types/profile/profile-layout';
+
+const getProblemsCnt = ({ allProblems, solvedProblems }: ProblemsCntType) => ({
+  allCnt: allProblems.length,
+  solvedCnt: solvedProblems.length,
+});
+
+export { getProblemsCnt };
